feat: Swagger for methods

+++ annotator-web/src/main/java/cz/cuni/mff/vopalenf/annotator/service/TeamService.java
@@ -2,6 +2,7 @@ package cz.cuni.mff.vopalenf.annotator.service;
 
 import cz.cuni.mff.vopalenf.annotator.api.model.Team;
 import cz.cuni.mff.vopalenf.annotator.dao.repository.TeamRepository;
+import cz.cuni.mff.vopalenf.annotator.exception.api.NotFoundException;
 import cz.cuni.mff.vopalenf.annotator.mapper.TeamMapper;
 import cz.cuni.mff.vopalenf.annotator.mapper.UserMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -27,6 +28,11 @@ public class TeamService {
         this.teamMapper = teamMapper;
     }
 
+    /**
+     * Get all teams in app
+     *
+     * @return List of all teams from database
+     */
     public List<Team> getAllTeams() {
         return teamRepository.findAll().stream()
                 .map(teamEntity -> teamMapper.mapTeam(
@@ -38,7 +44,16 @@ public class TeamService {
                 .toList();
     }
 
+    /**
+     * Delete team by its ID
+     *
+     * @param teamId ID of a team to delete
+     * @throws NotFoundException when such team ID does not exist
+     */
     public void deleteTeam(Long teamId) {
+        if (!teamRepository.existsById(teamId)) {
+            throw new NotFoundException("Not found team with id " + teamId, TeamService.class.getSimpleName());
+        }
         teamRepository.deleteById(teamId);
     }
 }
